frontend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
newspaper3k==0.2.8
lxml[html_clean]==5.3.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
